extract index reader, table reader, index look up reader out of "new_distsql.go
