+++ executor/index_reader.go
@@ -1,0 +1,141 @@
+// Copyright 2017 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package executor
+
+import (
+	"github.com/juju/errors"
+	"github.com/pingcap/tidb/context"
+	"github.com/pingcap/tidb/distsql"
+	"github.com/pingcap/tidb/expression"
+	"github.com/pingcap/tidb/model"
+	"github.com/pingcap/tidb/table"
+	"github.com/pingcap/tidb/terror"
+	"github.com/pingcap/tidb/util/types"
+	"github.com/pingcap/tipb/go-tipb"
+	goctx "golang.org/x/net/context"
+)
+
+// IndexReaderExecutor sends dag request and reads index data from kv layer.
+type IndexReaderExecutor struct {
+	table     table.Table
+	index     *model.IndexInfo
+	tableID   int64
+	keepOrder bool
+	desc      bool
+	ranges    []*types.IndexRange
+	dagPB     *tipb.DAGRequest
+	ctx       context.Context
+	schema    *expression.Schema
+
+	// result returns one or more distsql.PartialResult and each PartialResult is returned by one region.
+	result        distsql.NewSelectResult
+	partialResult distsql.NewPartialResult
+	// columns are only required by union scan.
+	columns  []*model.ColumnInfo
+	priority int
+}
+
+// Schema implements the Executor Schema interface.
+func (e *IndexReaderExecutor) Schema() *expression.Schema {
+	return e.schema
+}
+
+// Close implements the Executor Close interface.
+func (e *IndexReaderExecutor) Close() error {
+	err := closeAll(e.result, e.partialResult)
+	e.result = nil
+	e.partialResult = nil
+	return errors.Trace(err)
+}
+
+// Next implements the Executor Next interface.
+func (e *IndexReaderExecutor) Next() (Row, error) {
+	for {
+		// Get partial result.
+		if e.partialResult == nil {
+			var err error
+			e.partialResult, err = e.result.Next()
+			if err != nil {
+				return nil, errors.Trace(err)
+			}
+			if e.partialResult == nil {
+				// Finished.
+				return nil, nil
+			}
+		}
+		// Get a row from partial result.
+		rowData, err := e.partialResult.Next()
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		if rowData == nil {
+			// Finish the current partial result and get the next one.
+			err = e.partialResult.Close()
+			terror.Log(errors.Trace(err))
+			e.partialResult = nil
+			continue
+		}
+		err = decodeRawValues(rowData, e.schema, e.ctx.GetSessionVars().GetTimeZone())
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		return rowData, nil
+	}
+}
+
+// Open implements the Executor Open interface.
+func (e *IndexReaderExecutor) Open() error {
+	fieldTypes := make([]*types.FieldType, len(e.index.Columns))
+	for i, v := range e.index.Columns {
+		fieldTypes[i] = &(e.table.Cols()[v.Offset].FieldType)
+	}
+	var builder requestBuilder
+	kvReq, err := builder.SetIndexRanges(e.ctx.GetSessionVars().StmtCtx, e.tableID, e.index.ID, e.ranges, fieldTypes).
+		SetDAGRequest(e.dagPB).
+		SetDesc(e.desc).
+		SetKeepOrder(e.keepOrder).
+		SetPriority(e.priority).
+		SetFromSessionVars(e.ctx.GetSessionVars()).
+		Build()
+	if err != nil {
+		return errors.Trace(err)
+	}
+	e.result, err = distsql.NewSelectDAG(e.ctx.GoCtx(), e.ctx.GetClient(), kvReq, e.schema.Len())
+	if err != nil {
+		return errors.Trace(err)
+	}
+	e.result.Fetch(e.ctx.GoCtx())
+	return nil
+}
+
+// doRequestForDatums constructs kv ranges by datums. It is used by index look up executor.
+func (e *IndexReaderExecutor) doRequestForDatums(goCtx goctx.Context, values [][]types.Datum) error {
+	var builder requestBuilder
+	kvReq, err := builder.SetIndexValues(e.tableID, e.index.ID, values).
+		SetDAGRequest(e.dagPB).
+		SetDesc(e.desc).
+		SetKeepOrder(e.keepOrder).
+		SetPriority(e.priority).
+		SetFromSessionVars(e.ctx.GetSessionVars()).
+		Build()
+	if err != nil {
+		return errors.Trace(err)
+	}
+	e.result, err = distsql.NewSelectDAG(e.ctx.GoCtx(), e.ctx.GetClient(), kvReq, e.schema.Len())
+	if err != nil {
+		return errors.Trace(err)
+	}
+	e.result.Fetch(goCtx)
+	return nil
+}
